Add redirecting
Add the option to create a URL that, after going through the same login procedure, will redirect to the target URL specified. This is useful because sometimes due to the login procedure the original target URL (i.e. the URL of a specific course) is lost and only a general ILIAS page is reached. The extension can then manually redirect to the course page.